Refactor SingleLabelCreation component logic
